-- Mejoramos los codigos de respuesta en customer -- esto con el fin de hacer mas descriptivas las peticiones -- devolvemos un objeto json en lugar de una lista -- pero esta solucion no me deja muy satisfecho - a este punto debemos tener en cuenta la optimizacion ya que estamos usando varios hasmap y map
